+++ scripts/core-boundaries/self-test.mjs
@@ -2368,12 +2368,17 @@ export function runManifestParserSelfTest({
   ) {
     throw new Error('agentic system boundary rule must forbid terminal provider construction');
   }
-  const coreFileReadStateRuleText = forbiddenRuleTextForPath(
-    'src/crates/assembly/core/src/agentic/session/file_read_state.rs',
+  const coreReviewReadReceiptRuleText = forbiddenRuleTextForPath(
+    'src/crates/assembly/core/src/agentic/session/review_read_receipt.rs',
   );
-  for (const contract of ['FileReadState', 'FileReadStateStore', 'DashMap']) {
-    if (!coreFileReadStateRuleText.includes(contract)) {
-      throw new Error(`core file_read_state boundary rule must forbid ${contract}`);
+  for (const contract of [
+    'FileRevision',
+    'ReviewReadCoverage',
+    'ReviewReadReceiptStore',
+    'DashMap',
+  ]) {
+    if (!coreReviewReadReceiptRuleText.includes(contract)) {
+      throw new Error(`core review_read_receipt boundary rule must forbid ${contract}`);
     }
   }
   const coreEvidenceLedgerRuleText = forbiddenRuleTextForPath(
@@ -3180,13 +3185,13 @@ export function runManifestParserSelfTest({
       ],
     },
     {
-      path: 'src/crates/execution/agent-runtime/src/file_read_state.rs',
+      path: 'src/crates/execution/agent-runtime/src/review_read_receipt.rs',
       contracts: [
-        'FileReadState',
-        'is_full_file_read',
-        'FileReadStateStore',
-        'file_read_state_accepts_nonempty_whole_file',
-        'file_read_state_store_scopes_entries_by_session',
+        'FileRevision',
+        'ReviewReadCoverage',
+        'ReviewReadReceiptStore',
+        'review_read_receipt_store_scopes_entries_by_session',
+        'review_read_receipt_covers_only_previously_returned_lines',
       ],
     },
     {
@@ -3641,15 +3646,6 @@ export function runManifestParserSelfTest({
         'FILE_TOOL_GUIDANCE_PREFIX',
         'file_tool_guidance_message',
         'is_file_tool_guidance_message',
-      ],
-    },
-    {
-      path: 'src/crates/execution/tool-contracts/src/file_read_freshness.rs',
-      contracts: [
-        'FileReadFreshnessFacts',
-        'normalize_tool_file_content',
-        'file_read_facts_content_matches',
-        'file_read_facts_are_fresh',
       ],
     },
     {
