--- src/crates/assembly/core/src/agentic/session/file_read_state.rs
@@ -1,4 +1,0 @@
-//! Compatibility facade for session-scoped file read state.
-
-pub use openbitfun_agent_runtime::file_read_state::{FileReadState, FileReadStateStore};
-pub use openbitfun_agent_runtime::file_read_state::{FileRevision, ReviewReadCoverage};
