+++ src/crates/assembly/core/src/agentic/session/mod.rs
@@ -6,9 +6,9 @@ pub mod compression;
 pub mod context_store;
 mod context_usage;
 pub mod evidence_ledger;
-pub mod file_read_state;
 pub mod prompt_cache;
 pub(crate) mod revert;
+pub mod review_read_receipt;
 pub mod session_manager;
 pub mod session_store_port;
 pub mod token_anchor;
@@ -19,8 +19,8 @@ pub use compression::*;
 pub use context_store::*;
 pub use context_usage::*;
 pub use evidence_ledger::*;
-pub use file_read_state::*;
 pub use prompt_cache::*;
+pub use review_read_receipt::*;
 pub use session_manager::*;
 pub use session_store_port::*;
 pub use token_anchor::*;
