+++ src/crates/assembly/core/src/agentic/session/review_read_receipt.rs
@@ -1,0 +1,5 @@
+//! Compatibility facade for session-scoped code-review read receipts.
+
+pub use openbitfun_agent_runtime::review_read_receipt::{
+    FileRevision, ReviewReadCoverage, ReviewReadReceiptStore,
+};
