+++ src/crates/assembly/core/src/agentic/tools/mod.rs
@@ -7,7 +7,6 @@ pub mod computer_use_capability;
 pub mod computer_use_host;
 pub mod computer_use_optimizer;
 pub(crate) mod file_permissions;
-pub mod file_read_state_runtime;
 pub mod file_tool_guidance;
 pub mod framework;
 #[cfg(feature = "tools-creation")]
@@ -32,6 +31,7 @@ pub(crate) mod post_call_hooks;
 pub mod product_runtime;
 pub mod registry;
 pub mod restrictions;
+pub mod review_read_receipt_runtime;
 pub(crate) mod tool_adapter;
 pub(crate) mod tool_context_runtime;
 pub(crate) mod tool_result_storage;
